refactor: merged update user
